requirements_dev.txt
pip
bump2version
wheel
watchdog
flake8
tox
coverage
sphinx
sphinx_rtd_theme
twine
Click
pytest
black

numpy
scipy
matplotlib
xraylarch
